# Conflicts:
#	contracts/gas-snapshots/ccip.gas-snapshot
#	contracts/src/v0.8/ccip/test/pools/TokenPool/TokenPoolSetup.t.sol
#	contracts/src/v0.8/ccip/test/pools/USDC/HybridLockReleaseUSDCTokenPool/HybridLockReleaseUSDCTokenPool.lockOrBurn.t.sol
#	contracts/src/v0.8/ccip/test/pools/USDC/HybridLockReleaseUSDCTokenPool/HybridLockReleaseUSDCTokenPool.releaseOrMint.t.sol
#	contracts/src/v0.8/ccip/test/pools/USDC/HybridLockReleaseUSDCTokenPool/HybridLockReleaseUSDCTokenPool.transferLiquidity.t.sol
#	contracts/src/v0.8/ccip/test/pools/USDC/USDCBridgeMigrator/USDCBridgeMigrator.burnLockedUSDC.t.sol
#	contracts/src/v0.8/ccip/test/pools/USDC/USDCBridgeMigrator/USDCBridgeMigrator.cancelMigrationProposal.t.sol
#	contracts/src/v0.8/ccip/test/pools/USDC/USDCBridgeMigrator/USDCBridgeMigrator.excludeTokensFromBurn.t.sol
#	contracts/src/v0.8/ccip/test/pools/USDC/USDCBridgeMigrator/USDCBridgeMigrator.proposeMigration.t.sol
#	contracts/src/v0.8/ccip/test/pools/USDC/USDCBridgeMigrator/USDCBridgeMigrator.provideLiquidity.t.sol
#	contracts/src/v0.8/ccip/test/pools/USDC/USDCBridgeMigrator/USDCBridgeMigrator.releaseOrMint.t.sol
#	contracts/src/v0.8/ccip/test/pools/USDC/USDCBridgeMigrator/USDCBridgeMigrator.updateChainSelectorMechanism.t.sol

+++ integration-tests/ccip-tests/actions/reorg_helpers.go
@@ -27,16 +27,16 @@ type ReorgConfig struct {
 	// DstGethHTTPURL dest chain Geth HTTP URL
 	DstGethHTTPURL string
 	// SrcFinalityDepth source chain finality depth
-	SrcFinalityDepth uint64
+	SrcFinalityDepth int
 	// DstGethHTTPURL dest chain finality depth
-	DstFinalityDepth uint64
+	DstFinalityDepth int
 	// FinalityDelta blocks to rewind below or above finality
 	FinalityDelta int
 }
 
 // Validate validates ReorgConfig params
 func (rc *ReorgConfig) Validate() error {
-	if rc.FinalityDelta >= int(rc.SrcFinalityDepth) || rc.FinalityDelta >= int(rc.DstFinalityDepth) {
+	if rc.FinalityDelta >= rc.SrcFinalityDepth || rc.FinalityDelta >= rc.DstFinalityDepth {
 		return fmt.Errorf(
 			"finality delta can't be higher than source or dest chain finality, delta: %d, src: %d, dst: %d",
 			rc.FinalityDelta, rc.SrcFinalityDepth, rc.DstFinalityDepth,
